// src/infrastructure/api/dto/SwapiResponse.ts
export interface SwapiResponse<T> {
  count: number;
  next: string;
  previous: null | string;
  results: T[];
}

export interface SwapiPlanet {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
  created: Date;
  edited: Date;
  url: string;
}